src/main/java/loanbook/logic/commands/ExitCommand.java
package loanbook.logic.commands;

import loanbook.commons.core.EventsCenter;
import loanbook.commons.events.ui.ExitAppRequestEvent;
import loanbook.logic.CommandHistory;
import loanbook.model.Model;

/**
 * Terminates the program.
 */
public class ExitCommand extends Command {

    public static final String COMMAND_WORD = "exit";

    public static final String MESSAGE_EXIT_ACKNOWLEDGEMENT = "Exiting Loan Book as requested ...";

    @Override
    public CommandResult execute(Model model, CommandHistory history) {
        EventsCenter.getInstance().post(new ExitAppRequestEvent());
        return new CommandResult(MESSAGE_EXIT_ACKNOWLEDGEMENT);
    }

}
